Remove MatrixMap from RandomRoomGenerator
