+++ Assets/Scripts/Game/LevelManager/DungeonLoader/RandomRoom/MatrixMap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f9aeb1425e83aca4a885f2c15d598bdb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
